+++ src/chainlit/__init__.py
@@ -13,7 +13,7 @@ from chainlit.emitter import ChainlitEmitter
 from chainlit.types import LLMSettings
 from chainlit.message import ErrorMessage
 from chainlit.action import Action
-from chainlit.element import Image, Text, Pdf, Avatar, Pyplot
+from chainlit.element import Image, Text, Pdf, Avatar, Pyplot, Table
 from chainlit.message import Message, ErrorMessage, AskUserMessage, AskFileMessage
 from chainlit.user_session import user_session
 from chainlit.sync import run_sync, make_async
@@ -228,11 +228,12 @@ __all__ = [
     "user_session",
     "LLMSettings",
     "Action",
-    "Pdf",
-    "Image",
-    "Text",
     "Avatar",
+    "Image",
+    "Pdf",
     "Pyplot",
+    "Table",
+    "Text",
     "Message",
     "ErrorMessage",
     "AskUserMessage",